reorganizes cli code
